+++ src/main/java/com/cheocharm/MapZ/report/presentation/dto/ReportRequest.java
@@ -1,4 +1,4 @@
-package com.cheocharm.MapZ.report.domain.dto;
+package com.cheocharm.MapZ.report.presentation.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import javax.validation.constraints.NotNull;
 
 @Schema(description = "게시글 신고 Request Body")
 @Getter
-public class ReportDto {
+public class ReportRequest {
 
     @NotNull
     private Long diaryId;
